+++ examples/editor/src/App.jsx
@@ -16,6 +16,7 @@ import { testImages, testVideos } from './mock';
 import './App.css';
 import theme from './theme/theme'; // must import after custom styles
 import RichContentRawDataViewer from './RichContentRawDataViewer';
+import cloneDeep from 'lodash/cloneDeep'
 
 const modalStyleDefaults = {
   content: {
@@ -260,7 +261,7 @@ class App extends Component {
                   <div className="column side">
                     <RichContentRawDataViewer
                       onChange={content => this.setState({ content })}
-                      content={convertToRaw(this.state.editorState.getCurrentContent())}
+                      content={cloneDeep(convertToRaw(this.state.editorState.getCurrentContent()))}
                       width="740px"
                     />
                     <Button