+++ packages/plugin-map/src/MapViewer.jsx
@@ -1,0 +1,40 @@
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import { validate } from 'wix-rich-content-common';
+import isEqual from 'lodash/isEqual';
+import schema from '../statics/data-schema.json';
+import { Map } from './Map';
+
+export class MapViewer extends Component {
+  constructor(props) {
+    super(props);
+    validate(props.componentData, schema);
+  }
+
+  componentWillReceiveProps(nextProps) {
+    if (!isEqual(nextProps.componentData, this.props.componentData)) {
+      validate(nextProps.componentData, schema);
+    }
+  }
+
+  render() {
+    return <Map {...this.props} />;
+  }
+}
+
+MapViewer.propTypes = {
+  componentData: PropTypes.object.isRequired,
+  theme: PropTypes.object,
+  onReady: PropTypes.func,
+  onStart: PropTypes.func,
+  controls: PropTypes.bool,
+  width: PropTypes.string,
+  height: PropTypes.string,
+  settings: PropTypes.object,
+};
+
+MapViewer.defaultProps = {
+  width: '100%',
+  height: '100%',
+  controls: true,
+};