+++ packages/plugin-map/src/constants.js
@@ -1,0 +1,8 @@
+export const MAP_TYPE = 'wix-draft-plugin-map';
+
+export const DEFAULTS = {
+  config: {
+    size: 'content',
+    alignment: 'center',
+  },
+};