+++ packages/plugin-map/src/toolbar/insert-buttons.jsx
@@ -1,0 +1,17 @@
+import { TOOLBARS } from 'wix-rich-content-common';
+import { InsertPluginIcon } from '../icons';
+import { DEFAULTS } from '../constants';
+
+export default ({ helpers, t, settings }) => {
+  return [
+    {
+      name: 'Map',
+      tooltipText: t('MapPlugin_InsertButton_Tooltip'),
+      toolbars: [TOOLBARS.FOOTER, TOOLBARS.SIDE],
+      Icon: InsertPluginIcon,
+      componentData: Object.assign({}, DEFAULTS, settings),
+      helpers,
+      t,
+    },
+  ];
+};